% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirtfying data fns.R
\name{break.my_numbers}
\alias{break.my_numbers}
\title{Unclean a string column of a dataframe by adding some random errors}
\usage{
break.my_numbers(data.frame.column)
}
\description{
Developed with J Wilde for ASAB Spring 2024
}
