% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create edgelist from gbi.R
\name{create_edgelist}
\alias{create_edgelist}
\title{Create edgelist from a group by individual matrix}
\usage{
create_edgelist(gbi)
}
\description{
Create edgelist from a group by individual matrix
}
