% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirtfying data fns.R
\name{data_dirtifier}
\alias{data_dirtifier}
\title{Unclean a string column of a dataframe by adding some random errors}
\usage{
data_dirtifier(data_frame)
}
\description{
Developed with J Wilde for ASAB Spring 2024
}
